+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.42.3] - 2019-06-05
+### Fixed
+- Remove `showProductPrice` rule from `ProductPrice` wrapper.
+
 ## [3.42.2] - 2019-06-05
 
 ### Fixed